+++ WBLMS.IRepositories/IRepository.cs
@@ -4,12 +4,13 @@
     {
         Task<IEnumerable<T>> GetAsyncAll();
 
-        Task<T> GetAsyncById(long id);
+        Task<T> GetAsyncById(int id);
 
         Task<T> CreateAsync(T entity);
 
         Task<T> UpdateAsync(T entity);
 
         Task<bool> DeleteAsync(T entity);
+
     }
 }
